Renamed components in test
