+++ test/SearchContainer-test.js
@@ -4,7 +4,7 @@ import {assert} from 'chai'
 import React from 'react'
 import SearchContainer from '../app/containers/SearchContainer'
 import SuggestionsList from '../app/components/SuggestionsList'
-import IngredientSuggestion from '../app/components/IngredientSuggestion'
+import IngredientSuggestion from '../app/components/Ingredient'
 import SearchBar from '../app/components/SearchBar'
 import Preloader from '../app/components/Preloader'
 import ErrorMsg from '../app/components/ErrorMsg'
